+++ CHANGES.rst
@@ -4,7 +4,10 @@ Changelog for zest.releaser
 6.16.1 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Refuse to edit history header when it may be from an already released version.
+  We look for "unreleased" in it.  Give a warning when this happens.
+  Fixes `issue 311 <https://github.com/zestsoftware/zest.releaser/issues/311>`_.
+  [maurits]
 
 
 6.16.0 (2019-01-17)